+++ testapp/urls.py
@@ -19,5 +19,6 @@ urlpatterns = [
         name="settings",
     ),
     path("<slug:slug>", DetailView.as_view(model=TestModel), name="testmodel"),
+    path("<slug:slug>/update", views.TestUpdateView.as_view(), name="testmodel-update"),
     path("<slug:slug>/delete", views.TestDeleteView.as_view(), name="testmodel-del"),
 ]